move account details updates to AccountRegistrar
