+++ packages/client/src/layers/react/components/validators/WalletConnector.tsx
@@ -1,48 +1,17 @@
 import React, { useEffect, useState } from 'react';
-import { map, merge } from 'rxjs';
+import { of } from 'rxjs';
 import styled, { keyframes } from 'styled-components';
 import { useAccount, useNetwork, Connector } from 'wagmi';
-import {
-  EntityIndex,
-  Has,
-  HasValue,
-  getComponentValue,
-  runQuery,
-} from '@latticexyz/recs';
 
 import { defaultChainConfig } from 'constants/chains';
 import { createNetworkConfig } from 'layers/network/config';
 import { createNetworkLayer } from 'layers/network/createNetworkLayer';
-import { dataStore } from 'layers/react/store/createStore';
 import { useNetworkSettings } from 'layers/react/store/networkSettings';
 import { registerUIComponent } from 'layers/react/engine/store';
-import {
-  AccountDetails,
-  emptyAccountDetails,
-  useKamiAccount,
-} from 'layers/react/store/kamiAccount';
 import 'layers/react/styles/font.css';
 
-
-/** 
- * The sole purpose of this here monstrosity is to keep track of the connected Kami Account
- * based on the connected wallet address. Unfortunately, this means listening to both changes
- * in the Connector's address through State hooks, as well as to subscribed world components
- * on the Requirement step that may result in the creation of an account in-world.
- * 
- * The requirement step determines the Account's EntityIndex using a mirrored address saved on the
- * zustand store as wagmi's useAccount() is unavailable outside of React components. It is also
- * necessary to properly update the modal whenever the page is refreshed, causing a repopulation of
- * the world client-side.
- * 
- * The modal component then takes this index as a prop and simply listens to it. Nothing more. It
- * instead relies on a hook to the Same zustand store item for the Same connected account because
- * it's possible either side may be stale.
- * 
- * Let's not fool ourselves into thinking this is an elegant solution by any measure. It is an
- * abomination birthed out of necessity and should be treated as such.
- */
-
+// Detects network changes and populates network clients for inidividual addresses.
+// The purpose of this modal is to warn the user when something is amiss.
 export function registerWalletConnecter() {
   registerUIComponent(
     'WalletConnecter',
@@ -52,66 +21,8 @@ export function registerWalletConnecter() {
       rowStart: 40,
       rowEnd: 60,
     },
-    (layers) => {
-      const {
-        network: {
-          components: {
-            IsAccount,
-            Name,
-            OperatorAddress,
-            OwnerAddress,
-          },
-          network: { connectedAddress },
-          world,
-        },
-      } = layers;
-
-      const getAccountDetails = (index: EntityIndex): AccountDetails => {
-        if (!index) return emptyAccountDetails();
-        return {
-          id: world.entities[index],
-          index: index,
-          ownerAddress: getComponentValue(OwnerAddress, index)?.value as string,
-          operatorAddress: getComponentValue(OperatorAddress, index)?.value as string,
-          name: getComponentValue(Name, index)?.value as string,
-        };
-      }
-
-      const getAccountIndexFromOwner = (ownerAddress: string): EntityIndex => {
-        const accountIndex = Array.from(
-          runQuery([
-            Has(IsAccount),
-            HasValue(OwnerAddress, {
-              value: ownerAddress,
-            }),
-          ])
-        )[0];
-        return accountIndex;
-      };
-
-      return merge(
-        IsAccount.update$,
-        Name.update$,
-        OperatorAddress.update$,
-        OwnerAddress.update$,
-      ).pipe(
-        map(() => {
-          const { selectedAddress } = useNetworkSettings.getState();
-          const accountIndexUpdatedByWorld = getAccountIndexFromOwner(selectedAddress);
-          const accountDetailsFromWorld = getAccountDetails(accountIndexUpdatedByWorld);
-          return {
-            accountDetailsFromWorld,
-            getAccountIndexFromOwner,
-            getAccountDetails,
-          };
-        })
-      );
-    },
-    ({
-      accountDetailsFromWorld,
-      getAccountIndexFromOwner,
-      getAccountDetails,
-    }) => {
+    (layers) => of(layers),
+    () => {
       const { chain } = useNetwork();
 
       const {
@@ -124,53 +35,37 @@ export function registerWalletConnecter() {
       const {
         networks,
         addNetwork,
-        selectedAddress,
         setSelectedAddress,
       } = useNetworkSettings();
 
-      const { setDetails } = useKamiAccount();
-      const { toggleVisibleButtons, toggleVisibleModals } = dataStore();
       const [isCorrectNetwork, setIsCorrectNetwork] = useState(false);
       const [title, setTitle] = useState('Connect a Wallet');
       const [description, setDescription] = useState('');
 
       // check whether the correctNetwork is connected
+      // update title and description as needed
       useEffect(() => {
-        setIsCorrectNetwork(chain?.id === defaultChainConfig.id);
-      }, [isConnected, chain]);
-
-      // title and description as needed
-      useEffect(() => {
-        if (!isCorrectNetwork) {
-          setTitle('Wrong Network');
-          setDescription(`Please connect to ${defaultChainConfig.name}`);
-        } else {
+        const networksMatch = chain?.id === defaultChainConfig.id;
+        setIsCorrectNetwork(networksMatch);
+        if (!isConnected) {
           setTitle('Connect a Wallet');
           setDescription('You must connect a wallet to continue.');
+        } else if (!networksMatch) {
+          setTitle('Wrong Network');
+          setDescription(`Please connect to ${defaultChainConfig.name}`);
         }
-      }, [isCorrectNetwork]);
-
-      // track the account details in store for easy access
-      // expose/hide components accordingly
-      useEffect(() => {
-        const accountIndex = getAccountIndexFromOwner(selectedAddress);
-        const accountDetails = getAccountDetails(accountIndex);
-        setDetails(accountDetails);
-
-        if (accountDetails.id) {
-          toggleVisibleButtons(true);
-        } else {
-          toggleVisibleButtons(false);
-          toggleVisibleModals(false);
-        }
-      }, [selectedAddress, isConnected, accountDetailsFromWorld]);
+      }, [isConnected, chain]);
 
       // update the network settings whenever the connector/address changes
       useEffect(() => {
         console.log("WALLET IS", status);
         console.log("NETWORK CHANGE DETECTED");
         updateNetworkSettings(connector);
-      }, [chain, connector, connectorAddress, isConnected]);
+      }, [chain, connector, connectorAddress, isConnected, isCorrectNetwork]);
+
+
+      /////////////////
+      // ACTIONS
 
       // add a network layer if one for the connection doesnt exist
       const updateNetworkSettings = async (connector: Connector | undefined) => {
@@ -198,6 +93,7 @@ export function registerWalletConnecter() {
         }
       };
 
+
       /////////////////
       // RENDER
 
@@ -205,7 +101,6 @@ export function registerWalletConnecter() {
       const modalDisplay = () => (
         (isConnected && isCorrectNetwork) ? 'none' : 'block'
       );
-
 
       return (
         <ModalWrapper id='connect' style={{ display: modalDisplay() }}>
@@ -225,13 +120,14 @@ const Title = styled.p`
   color: #333;
   text-align: center;
   font-family: Pixel;
+  padding: 10px;
 `;
 
 const Description = styled.p`
   font-size: 12px;
   color: #333;
   text-align: center;
-  padding: 10px 0px 15px 0px;
+  padding: 0px 0px 20px 0px;
   font-family: Pixel;
 `;
